test(app): keep count contracts with session inventory coverage
